airflow_src/requirements.txt
pandas
boto3==1.13.26
botocore==1.16.26
requests==2.23.0
snowflake-sqlalchemy==1.2.3
snowflake-connector-python==2.3.7
idna==2.7
cryptography==3.2.0
pytz==2020.5